+++ src/models/parameters.ts
@@ -10,6 +10,8 @@ import {
   CollectionCreate as CollectionCreateMapper,
   CollectionUpdate as CollectionUpdateMapper,
   DocumentUpdate as DocumentUpdateMapper,
+  BatchAddTagRequest as BatchAddTagRequestMapper,
+  BatchRemoveTagRequest as BatchRemoveTagRequestMapper,
   DocumentEditRequest as DocumentEditRequestMapper,
   ExtractorCreate as ExtractorCreateMapper,
   ExtractorUpdate as ExtractorUpdateMapper,
@@ -523,6 +525,16 @@ export const body5: OperationParameter = {
 
 export const body6: OperationParameter = {
   parameterPath: "body",
+  mapper: BatchAddTagRequestMapper
+};
+
+export const body7: OperationParameter = {
+  parameterPath: "body",
+  mapper: BatchRemoveTagRequestMapper
+};
+
+export const body8: OperationParameter = {
+  parameterPath: "body",
   mapper: DocumentEditRequestMapper
 };
 
@@ -537,12 +549,12 @@ export const includePublicExtractors: OperationQueryParameter = {
   }
 };
 
-export const body7: OperationParameter = {
+export const body9: OperationParameter = {
   parameterPath: ["options", "body"],
   mapper: ExtractorCreateMapper
 };
 
-export const body8: OperationParameter = {
+export const body10: OperationParameter = {
   parameterPath: "body",
   mapper: ExtractorUpdateMapper
 };
@@ -619,12 +631,12 @@ export const identifier2: OperationQueryParameter = {
   }
 };
 
-export const body9: OperationParameter = {
+export const body11: OperationParameter = {
   parameterPath: ["options", "body"],
   mapper: DataPointCreateMapper
 };
 
-export const body10: OperationParameter = {
+export const body12: OperationParameter = {
   parameterPath: "body",
   mapper: DataPointUpdateMapper
 };
@@ -653,7 +665,7 @@ export const collection2: OperationQueryParameter = {
   }
 };
 
-export const body11: OperationParameter = {
+export const body13: OperationParameter = {
   parameterPath: ["options", "body"],
   mapper: DataPointChoiceCreateMapper
 };
@@ -670,7 +682,7 @@ export const id: OperationURLParameter = {
   }
 };
 
-export const body12: OperationParameter = {
+export const body14: OperationParameter = {
   parameterPath: "body",
   mapper: DataPointChoiceUpdateMapper
 };
@@ -687,17 +699,17 @@ export const document: OperationQueryParameter = {
   }
 };
 
-export const body13: OperationParameter = {
+export const body15: OperationParameter = {
   parameterPath: "body",
   mapper: AnnotationCreateMapper
 };
 
-export const body14: OperationParameter = {
+export const body16: OperationParameter = {
   parameterPath: "body",
   mapper: AnnotationUpdateMapper
 };
 
-export const body15: OperationParameter = {
+export const body17: OperationParameter = {
   parameterPath: "body",
   mapper: {
     serializedName: "body",
@@ -716,7 +728,7 @@ export const body15: OperationParameter = {
   }
 };
 
-export const body16: OperationParameter = {
+export const body18: OperationParameter = {
   parameterPath: "body",
   mapper: {
     serializedName: "body",
@@ -735,7 +747,7 @@ export const body16: OperationParameter = {
   }
 };
 
-export const body17: OperationParameter = {
+export const body19: OperationParameter = {
   parameterPath: "body",
   mapper: {
     serializedName: "body",
@@ -753,12 +765,12 @@ export const body17: OperationParameter = {
   }
 };
 
-export const body18: OperationParameter = {
+export const body20: OperationParameter = {
   parameterPath: "body",
   mapper: TagCreateMapper
 };
 
-export const body19: OperationParameter = {
+export const body21: OperationParameter = {
   parameterPath: "body",
   mapper: TagUpdateMapper
 };
@@ -824,7 +836,7 @@ export const role: OperationQueryParameter = {
   }
 };
 
-export const body20: OperationParameter = {
+export const body22: OperationParameter = {
   parameterPath: "body",
   mapper: OrganizationMembershipUpdateMapper
 };
@@ -840,12 +852,12 @@ export const status: OperationQueryParameter = {
   }
 };
 
-export const body21: OperationParameter = {
+export const body23: OperationParameter = {
   parameterPath: "body",
   mapper: InvitationCreateMapper
 };
 
-export const body22: OperationParameter = {
+export const body24: OperationParameter = {
   parameterPath: "body",
   mapper: InvitationUpdateMapper
 };
@@ -862,17 +874,17 @@ export const token: OperationURLParameter = {
   }
 };
 
-export const body23: OperationParameter = {
+export const body25: OperationParameter = {
   parameterPath: "body",
   mapper: InvitationResponseMapper
 };
 
-export const body24: OperationParameter = {
+export const body26: OperationParameter = {
   parameterPath: "body",
   mapper: ResthookSubscriptionCreateMapper
 };
 
-export const body25: OperationParameter = {
+export const body27: OperationParameter = {
   parameterPath: "body",
   mapper: ResthookSubscriptionUpdateMapper
 };
@@ -889,17 +901,17 @@ export const xHookSecret: OperationParameter = {
   }
 };
 
-export const body26: OperationParameter = {
+export const body28: OperationParameter = {
   parameterPath: "body",
   mapper: JobDescriptionSearchParametersMapper
 };
 
-export const body27: OperationParameter = {
+export const body29: OperationParameter = {
   parameterPath: "body",
   mapper: JobDescriptionSearchConfigMapper
 };
 
-export const body28: OperationParameter = {
+export const body30: OperationParameter = {
   parameterPath: ["options", "body"],
   mapper: PathsM3DzbgV3JobDescriptionSearchEmbedPostRequestbodyContentApplicationJsonSchemaMapper
 };
@@ -938,12 +950,12 @@ export const name3: OperationURLParameter = {
   }
 };
 
-export const body29: OperationParameter = {
+export const body31: OperationParameter = {
   parameterPath: "body",
   mapper: PathsCl024WV3IndexNameDocumentsPostRequestbodyContentApplicationJsonSchemaMapper
 };
 
-export const body30: OperationParameter = {
+export const body32: OperationParameter = {
   parameterPath: "body",
   mapper: ResumeSearchParametersMapper
 };
@@ -1129,12 +1141,12 @@ export const managementLevelWeight: OperationQueryParameter = {
   }
 };
 
-export const body31: OperationParameter = {
+export const body33: OperationParameter = {
   parameterPath: "body",
   mapper: ResumeSearchConfigMapper
 };
 
-export const body32: OperationParameter = {
+export const body34: OperationParameter = {
   parameterPath: ["options", "body"],
   mapper: Paths1Czpnk1V3ResumeSearchEmbedPostRequestbodyContentApplicationJsonSchemaMapper
 };