+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Update python_requires to be PEP compliant
 
+### Added
+- Endpoints for add/remove tag for documents
+
 ## [6.9.0] - 2023-05-02
 ### Added
 - Add custom data to job description search results